requirements_minimal.txt
# Minimal requirements for JARVIS to run
fastapi
uvicorn[standard]
pydantic
python-multipart
websockets
aiohttp
requests
python-dotenv
psutil
objgraph
pympler
numpy
scipy
langchain
langchain-community
spacy
transformers
torch
llama-cpp-python